requirements.txt
fastapi==0.82.0
pydantic==1.10.2
pydantic[email]
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.18.3